+++ Controllers/ReviewController.cs
@@ -1,4 +1,5 @@
-﻿using ActiveX_Api.Dto.Review;
+﻿using ActiveX_Api.Constants;
+using ActiveX_Api.Dto.Review;
 using ActiveX_Api.Mappers;
 using ActiveX_Api.Models;
 using Microsoft.AspNetCore.Authorization;
@@ -33,7 +34,10 @@ namespace ActiveX_Api.Controllers
             var product = await _context.Products.FindAsync(reviewDto.ProductId);
             if (product == null) return BadRequest("Invalid ProductId");
 
-            var review = reviewDto.FromCreateReviewDtoToReview(); 
+            var review = reviewDto.FromCreateReviewDtoToReview();
+            var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimNames.Id)?.Value;
+            review.UserId = userId;
+
             await _context.Reviews.AddAsync(review);
             await _context.SaveChangesAsync();
 
@@ -45,6 +49,9 @@ namespace ActiveX_Api.Controllers
         {
             var review = await _context.Reviews.FindAsync(id);
             if (review == null) return NotFound();
+
+            var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimNames.Id)?.Value;
+            if (review.UserId != userId) return Unauthorized("Only owner can edit review");
 
             review.Title = reviewDto.Title;
             review.Content = reviewDto.Content;
@@ -60,6 +67,9 @@ namespace ActiveX_Api.Controllers
         {
             var review = await _context.Reviews.FindAsync(id);
             if (review == null) return NotFound();
+
+            var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimNames.Id)?.Value;
+            if (review.UserId != userId) return Unauthorized("Only owner can delete review");
 
             _context.Reviews.Remove(review);
             await _context.SaveChangesAsync();
